ivy: get rid of historian and start using AMX
I discovered AMX thanks to Protesilaos Stavrou's helpful video in Ivy
configuration (https://www.youtube.com/watch?v=QcPenVmmQyM&t=483s).
And it appears to be exactly what I was looking for. This is a major
functionality of Helm that I missed when I made the switch to Ivy.
Historian was not good enough. This is great out of the box.