config: replace ivy with selectrum and extras
A quick glance at the documentation for these packages indicates that
this is a better option. It relies on manipulating default Emacs
interfaces to provide a selection mechanism.

There are cons to this approach, especially given that Ivy and Helm
are more widely supported. But I'm presuming this con is outweighted
by the fact that I can write custom functions to replace some Ivy
features.