org: use :unnarrowed for capturing interruptionse, meetings & calls
There is a condition in `org-capture.el` that checks whether the
currently running clock is less than `point-max`, and it fails if the
buffer is narrowed. Till we understand why the condition exists and
what purpose it serves, let's use unnarrowed buffers for capturing
these so that the previous clock is resumed properly.