evil: major overhaul again bringing evil mode back into the fold
I thought long and hard about this change and ultimately decided it is
for the best. With my Iris keyboard for example, I am having to
scrunch my fingers in weird ways to hit some of the basic Emacs keys.

With this commit, we're also bringing General back into the fold and
configuring all of my keys to be in common with both Emacs and Evil
modes, should I ever choose to leave Evil mode again.