org: ungroup @home and @work tags
I should ideally rename this to HOME and WORK instead of using the @
symbol, which is normally used to indicate a context. But this will
have to do for now.

This means I can mark some work related tasks also as home so that I
can get it done there also.